refactor:getContents 메서드 수정
