+++ src/main/kotlin/com/dooingle/domain/catchdomain/dto/DeleteCatchRequest.kt
@@ -1,4 +1,4 @@
-package com.dooingle.domain.catch.dto
+package com.dooingle.domain.catchdomain.dto
 
 // TODO : 추후 로그인 기능이 구현되면 삭제할 DTO
 data class DeleteCatchRequest(
