+++ src/main/kotlin/com/dooingle/domain/dooingle/dto/DooingleAndCatchResponse.kt
@@ -1,11 +1,12 @@
 package com.dooingle.domain.dooingle.dto
 
+import com.dooingle.domain.catchdomain.dto.CatchResponse
 import java.time.ZonedDateTime
 
 data class DooingleAndCatchResponse(
     val ownerName: String,
     val dooingleId: Long,
     val content: String,
-    val catch: Any?,
+    val catch: CatchResponse?,
     val createdAt: ZonedDateTime
 )
