+++ src/main/kotlin/com/dooingle/domain/dooingle/repository/DooingleQueryDslRepository.kt
@@ -1,5 +1,6 @@
 package com.dooingle.domain.dooingle.repository
 
+import com.dooingle.domain.dooingle.dto.DooingleAndCatchResponse
 import com.dooingle.domain.dooingle.dto.DooingleResponse
 import com.dooingle.domain.user.model.SocialUser
 import org.springframework.data.domain.Pageable
@@ -9,7 +10,7 @@ interface DooingleQueryDslRepository {
 
     fun getDooinglesBySlice(cursor: Long?, pageable: Pageable): Slice<DooingleResponse>
 
-    fun getPersonalPageBySlice(owner: SocialUser, cursor: Long?, pageable: Pageable): Slice<DooingleResponse>
+    fun getPersonalPageBySlice(owner: SocialUser, cursor: Long?, pageable: Pageable): Slice<DooingleAndCatchResponse>
 
     // TODO 팔로우 기능 구현 후 구현 필요
     // fun getDooinglePageableOfFollows(pageable: Pageable): Slice<Dooingle>
